task_queue/requirements.txt
boto3==1.24.32
botocore==1.27.32
certifi==2022.6.15
charset-normalizer==2.1.0
dnspython==2.2.1
idna==3.3
jmespath==1.0.1
loguru==0.6.0
numpy==1.23.1
pinecone-client==2.0.11
pyarrow==8.0.0
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.16.6
s3transfer==0.6.0
six==1.16.0
tabulate==0.8.10
towhee==0.7.2
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.10
